Add X-Forwarded-For patch from http://haproxy.1wt.eu/download/patches/stunnel-4.15-xforwarded-for.diff
This patch for stunnel-4.15 adds support for the X-Forwarded-For header
insertion when the option 'xforwardedfor' is set to 'yes'. Unlike some
other patches found on the net, this one does reserve some space for the
insertion so it does not introduce any buffer overflow. Moreover, it
inserts the header after all other headers and not at the top, which
respects the order of headers even if the client sets the same header.

   Willy Tarreau